+++ core/src/main/java/com/graphhopper/GraphHopper.java
@@ -112,7 +112,7 @@ public class GraphHopper {
     private double residentialAreaSensitivity = 60;
     private double cityAreaRadius = 2000;
     private double cityAreaSensitivity = 30;
-    private int developmentCalculationThreads = 0;
+    private int urbanDensityCalculationThreads = 0;
 
     // preparation handlers
     private final LMPreparationHandler lmPreparationHandler = new LMPreparationHandler();
@@ -195,7 +195,7 @@ public class GraphHopper {
     }
 
     /**
-     * Configures the development classification. Each edge will be classified as 'rural','residential' or 'city', {@link Development}
+     * Configures the urban density classification. Each edge will be classified as 'rural','residential' or 'city', {@link UrbanDensity}
      *
      * @param residentialAreaRadius      in meters. The higher this value the longer the calculation will take and the bigger the area for
      *                                   which the road density used to identify residential areas is calculated.
@@ -206,17 +206,17 @@ public class GraphHopper {
      *                                   to skip the city classification.
      * @param cityAreaSensitivity        Use this to find a trade-off between too many roads being classified as city (too high values)
      *                                   and not enough roads being classified as city (too small values)
-     * @param threads                    the number of threads used for the calculation. If this is zero the development
-     *                                   classification is skipped entirely
+     * @param threads                    the number of threads used for the calculation. If this is zero the urban density
+     *                                   calculation is skipped entirely
      */
-    public GraphHopper setDevelopmentClassification(double residentialAreaRadius, double residentialAreaSensitivity,
-                                                    double cityAreaRadius, double cityAreaSensitivity, int threads) {
+    public GraphHopper setUrbanDensityCalculation(double residentialAreaRadius, double residentialAreaSensitivity,
+                                                  double cityAreaRadius, double cityAreaSensitivity, int threads) {
         ensureNotLoaded();
         this.residentialAreaRadius = residentialAreaRadius;
         this.residentialAreaSensitivity = residentialAreaSensitivity;
         this.cityAreaRadius = cityAreaRadius;
         this.cityAreaSensitivity = cityAreaSensitivity;
-        this.developmentCalculationThreads = threads;
+        this.urbanDensityCalculationThreads = threads;
         return this;
     }
 
@@ -571,12 +571,12 @@ public class GraphHopper {
         preciseIndexResolution = ghConfig.getInt("index.high_resolution", preciseIndexResolution);
         maxRegionSearch = ghConfig.getInt("index.max_region_search", maxRegionSearch);
 
-        // development calculation
-        residentialAreaRadius = ghConfig.getDouble("graph.development.residential_radius", residentialAreaRadius);
-        residentialAreaSensitivity = ghConfig.getDouble("graph.development.residential_sensitivity", residentialAreaSensitivity);
-        cityAreaRadius = ghConfig.getDouble("graph.development.city_radius", cityAreaRadius);
-        cityAreaSensitivity = ghConfig.getDouble("graph.development.city_sensitivity", cityAreaSensitivity);
-        developmentCalculationThreads = ghConfig.getInt("graph.development.threads", developmentCalculationThreads);
+        // urban density calculation
+        residentialAreaRadius = ghConfig.getDouble("graph.urban_density.residential_radius", residentialAreaRadius);
+        residentialAreaSensitivity = ghConfig.getDouble("graph.urban_density.residential_sensitivity", residentialAreaSensitivity);
+        cityAreaRadius = ghConfig.getDouble("graph.urban_density.city_radius", cityAreaRadius);
+        cityAreaSensitivity = ghConfig.getDouble("graph.urban_density.city_sensitivity", cityAreaSensitivity);
+        urbanDensityCalculationThreads = ghConfig.getInt("graph.urban_density.threads", urbanDensityCalculationThreads);
 
         // routing
         routerConfig.setMaxVisitedNodes(ghConfig.getInt(Routing.INIT_MAX_VISITED_NODES, routerConfig.getMaxVisitedNodes()));
@@ -592,7 +592,7 @@ public class GraphHopper {
         return this;
     }
 
-    private void buildEncodingManagerAndOSMParsers(String flagEncodersStr, String encodedValuesStr, String dateRangeParserString, boolean withDevelopment, Collection<Profile> profiles) {
+    private void buildEncodingManagerAndOSMParsers(String flagEncodersStr, String encodedValuesStr, String dateRangeParserString, boolean withUrbanDensity, Collection<Profile> profiles) {
         Map<String, String> flagEncodersMap = new LinkedHashMap<>();
         for (String encoderStr : flagEncodersStr.split(",")) {
             String name = encoderStr.split("\\|")[0].trim();
@@ -620,8 +620,8 @@ public class GraphHopper {
         EncodingManager.Builder emBuilder = new EncodingManager.Builder();
         flagEncodersMap.forEach((name, encoderStr) -> emBuilder.add(vehicleEncodedValuesFactory.createVehicleEncodedValues(name, new PMap(encoderStr))));
         profiles.forEach(profile -> emBuilder.add(Subnetwork.create(profile.getName())));
-        if (withDevelopment)
-            emBuilder.add(Development.create());
+        if (withUrbanDensity)
+            emBuilder.add(UrbanDensity.create());
         encodedValueStrings.forEach(s -> emBuilder.add(encodedValueFactory.create(s)));
         encodingManager = emBuilder.build();
 
@@ -780,8 +780,8 @@ public class GraphHopper {
     private void process(boolean closeEarly) {
         GHDirectory directory = new GHDirectory(ghLocation, dataAccessDefaultType);
         directory.configure(dataAccessConfig);
-        boolean withDevelopment = developmentCalculationThreads > 0;
-        buildEncodingManagerAndOSMParsers(flagEncodersString, encodedValuesString, dateRangeParserString, withDevelopment, profilesByName.values());
+        boolean withUrbanDensity = urbanDensityCalculationThreads > 0;
+        buildEncodingManagerAndOSMParsers(flagEncodersString, encodedValuesString, dateRangeParserString, withUrbanDensity, profilesByName.values());
         baseGraph = new BaseGraph.Builder(getEncodingManager())
                 .setDir(directory)
                 .set3D(hasElevation())
@@ -822,15 +822,15 @@ public class GraphHopper {
         if (hasElevation())
             interpolateBridgesTunnelsAndFerries();
 
-        if (encodingManager.hasEncodedValue(Development.KEY)) {
-            EnumEncodedValue<Development> developmentEnc = encodingManager.getEnumEncodedValue(Development.KEY, Development.class);
+        if (encodingManager.hasEncodedValue(UrbanDensity.KEY)) {
+            EnumEncodedValue<UrbanDensity> urbanDensityEnc = encodingManager.getEnumEncodedValue(UrbanDensity.KEY, UrbanDensity.class);
             if (!encodingManager.hasEncodedValue(RoadClass.KEY))
-                throw new IllegalArgumentException("Development calculation requires " + RoadClass.KEY);
+                throw new IllegalArgumentException("Urban density calculation requires " + RoadClass.KEY);
             if (!encodingManager.hasEncodedValue(RoadClassLink.KEY))
-                throw new IllegalArgumentException("Development calculation requires " + RoadClassLink.KEY);
+                throw new IllegalArgumentException("Urban density calculation requires " + RoadClassLink.KEY);
             EnumEncodedValue<RoadClass> roadClassEnc = encodingManager.getEnumEncodedValue(RoadClass.KEY, RoadClass.class);
             BooleanEncodedValue roadClassLinkEnc = encodingManager.getBooleanEncodedValue(RoadClassLink.KEY);
-            DevelopmentCalculator.calcDevelopment(baseGraph, developmentEnc, roadClassEnc, roadClassLinkEnc, residentialAreaRadius, residentialAreaSensitivity, cityAreaRadius, cityAreaSensitivity, developmentCalculationThreads);
+            UrbanDensityCalculator.calcUrbanDensity(baseGraph, urbanDensityEnc, roadClassEnc, roadClassLinkEnc, residentialAreaRadius, residentialAreaSensitivity, cityAreaRadius, cityAreaSensitivity, urbanDensityCalculationThreads);
         }
     }